Begin to implement unit tests using ERT
I said that I would do this a looooong time ago, but am only now
getting around to it (as a result of feeling envigorated to write
code lately).  This patch implements some basic unit tests for our
`php-runtime` library, and in particular the `php-runtime-expr`
part of the API.

The unit test code provides a `run-simple-php` function to help with
testing, allowing us to hand some short code to the PHP executable
for execution.  However, the `php-runtime` library provides this exact
functionality---obviously it must.  Therefore in the future we should
remove our utility from the unit test code and instead rely on the
relevant code from the `php-runtime` library itself.  The only reason
I did not use it in this first draft is because I wanted to re-examine
the code to see if I can simplify it in any meaningful and useful way.

Finally, this file of unit tests is incomplete and we will be adding
more tests in the future, particularly for such things as reading
standard-input into PHP code we wish to run.

GitHub-Issue: #2
Signed-off-by: Eric James Michael Ritz <ejmr@no.current.address>